injection new classe
